requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-compressor==2.4
django-libsass==0.8
libsass==0.20.1
Pillow==8.1.2
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
django-ckeditor
